modelx/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional: For ONNX to TensorFlow.js conversion
# tensorflowjs>=4.0.0
